feature: allow to specify warmup seconds using decimal format.

+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/teleport_warmup/config/model/TeleportWarmupConfigModel.java
@@ -12,7 +12,7 @@ public class TeleportWarmupConfigModel {
     @Document(id = 1751826785758L, value = """
         The `warmup seconds` for `all` teleports.
         """)
-    public int warmup_second = 3;
+    public double warmup_second = 3;
 
     public Interruptible interruptible = new Interruptible(true, 1, true, true);
